refactor(channelui): hoist HighlightMentions, thread-render leaves, SummarizeUnreadMessages
Three small pure clusters off cmd/wuphf and into channelui:

mentions.go:
- HighlightMentions wraps every "@slug" in a colored bold style
  based on a slug-to-color map. The private mentionPattern regex
  came along.

thread_render.go:
- FlattenThreadReplies — depth-first walk of descendants, tagged
  with depth + parent label
- RenderThreadReplies / RenderThreadReply — per-reply header+body
  rendering
- RenderThreadMessage — compact parent-style layout

The channelModel-unbound but tui-dependent renderThreadPanel and
renderThreadInput stay in package main since channelui does not
take a tui dependency yet.

unread.go:
- SummarizeUnreadMessages renders "N new from <names>" naming up
  to three distinct senders.

Aliases preserve every existing lowercase callsite — channel.go,
channel_window.go, channel_thread.go, the render-extras test all
keep compiling unchanged.
