+++ cmd/wuphf/channel_thread.go
@@ -149,144 +149,6 @@ func renderThreadPanel(allMessages []brokerMessage, parentID string,
 	return strings.Join(rendered, "\n")
 }
 
-func flattenThreadReplies(allMessages []brokerMessage, parentID string) []threadedMessage {
-	byID := make(map[string]brokerMessage, len(allMessages))
-	children := buildReplyChildren(allMessages)
-	for _, msg := range allMessages {
-		byID[msg.ID] = msg
-	}
-
-	var out []threadedMessage
-	var walk func(id string, depth int)
-	walk = func(id string, depth int) {
-		for _, child := range children[id] {
-			parentLabel := parentID
-			if parent, ok := byID[child.ReplyTo]; ok {
-				parentLabel = "@" + parent.From
-			}
-			out = append(out, threadedMessage{
-				Message:     child,
-				Depth:       depth,
-				ParentLabel: parentLabel,
-			})
-			walk(child.ID, depth+1)
-		}
-	}
-
-	walk(parentID, 0)
-	return out
-}
-
-func renderThreadReplies(replies []threadedMessage, width int) []string {
-	if len(replies) == 0 {
-		return nil
-	}
-
-	var lines []string
-	for _, reply := range replies {
-		lines = append(lines, renderThreadReply(reply, width)...)
-	}
-	return lines
-}
-
-func renderThreadReply(reply threadedMessage, width int) []string {
-	msg := reply.Message
-	color := agentColorMap[msg.From]
-	if color == "" {
-		color = "#9CA3AF"
-	}
-	nameStyle := lipgloss.NewStyle().
-		Foreground(lipgloss.Color(color)).
-		Bold(true)
-	tsStyle := lipgloss.NewStyle().
-		Foreground(lipgloss.Color(slackTimestamp))
-	metaStyle := lipgloss.NewStyle().
-		Foreground(lipgloss.Color(slackMuted))
-
-	name := displayName(msg.From)
-	ts := formatShortTime(msg.Timestamp)
-
-	prefix := "  " + strings.Repeat("  ", reply.Depth)
-	if reply.Depth > 0 {
-		prefix += "↳ "
-	}
-
-	meta := roleLabel(msg.From)
-	if usageMeta := renderMessageUsageMeta(msg.Usage, color); usageMeta != "" {
-		meta += " · " + usageMeta
-	}
-	if reply.ParentLabel != "" {
-		meta += " · reply to " + reply.ParentLabel
-	}
-
-	var lines []string
-	lines = append(lines, fmt.Sprintf("%s%s %s  %s  %s",
-		prefix,
-		agentAvatar(msg.From),
-		nameStyle.Render(name),
-		tsStyle.Render(ts),
-		metaStyle.Render(meta),
-	))
-
-	bodyPrefix := "  " + strings.Repeat("  ", reply.Depth)
-	if reply.Depth > 0 {
-		bodyPrefix += lipgloss.NewStyle().Foreground(lipgloss.Color(color)).Render("┆") + " "
-	} else {
-		bodyPrefix += lipgloss.NewStyle().Foreground(lipgloss.Color(color)).Render("│") + " "
-	}
-
-	for _, paragraph := range strings.Split(msg.Content, "\n") {
-		paragraph = highlightMentions(paragraph, agentColorMap)
-		for _, wrappedLine := range strings.Split(ansi.Wrap(paragraph, width-4, ""), "\n") {
-			lines = append(lines, bodyPrefix+wrappedLine)
-		}
-	}
-	lines = append(lines, "")
-	return lines
-}
-
-// renderThreadMessage renders a single message in thread style (compact).
-func renderThreadMessage(msg brokerMessage, width int, isParent bool) []string {
-	color := agentColorMap[msg.From]
-	if color == "" {
-		color = "#9CA3AF"
-	}
-	nameStyle := lipgloss.NewStyle().
-		Foreground(lipgloss.Color(color)).
-		Bold(true)
-	tsStyle := lipgloss.NewStyle().
-		Foreground(lipgloss.Color(slackTimestamp))
-
-	name := displayName(msg.From)
-	ts := formatShortTime(msg.Timestamp)
-
-	nameRendered := nameStyle.Render(name)
-	tsRendered := tsStyle.Render(ts)
-	nameWidth := lipgloss.Width(nameRendered)
-	tsWidth := lipgloss.Width(tsRendered)
-	gap := width - nameWidth - tsWidth - 4
-	if gap < 2 {
-		gap = 2
-	}
-
-	var lines []string
-	lines = append(lines, fmt.Sprintf("  %s %s%s%s",
-		agentAvatar(msg.From), nameRendered, strings.Repeat(" ", gap), tsRendered))
-	if usageMeta := renderMessageUsageMeta(msg.Usage, color); usageMeta != "" {
-		lines = append(lines, "  "+usageMeta)
-	}
-
-	// Render content
-	for _, paragraph := range strings.Split(msg.Content, "\n") {
-		paragraph = highlightMentions(paragraph, agentColorMap)
-		wrapped := ansi.Wrap(paragraph, width-4, "")
-		for _, wl := range strings.Split(wrapped, "\n") {
-			lines = append(lines, "  "+wl)
-		}
-	}
-	return lines
-}
-
 // renderThreadInput renders the input area at the bottom of the thread panel.
 func renderThreadInput(input []rune, inputPos int, width int, focused bool, historyAvailable bool) string {
 	if width < 6 {
